Move canvas size out of draw and draw outline
Setting the canvas size every time would allow for the canvas to fit to
the size of the window, but sometimes this isn't wanted. It also makes
more sense for this to be done in update, not draw. Since the size is
no longer changing, the canvas has to be cleared every time.

The outline was added to help distinguish between the canvas and the
background.